pyproject.toml
[tool.poetry]
name = "external-media"
version = "0.1.1"
description = ""
authors = ["EliseyGodX"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
python-ulid = "^3.0.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-isort = "^6.1.1"
flake8-bugbear = "^24.12.12"
flake8-pyproject = "^1.2.3"
flake8-docstrings = "^1.7.0"
black = "^24.10.0"

[tool.black]
line-length = 110

[tool.flake8]
max-line-length = 110
required-plugins = [
    "flake8-annotations",
    "flake8-bugbear",
    "flake8-docstrings",
    "flake8-isort",
    "flake8-pyproject",
    "mccabe",
    "pycodestyle",
    "pyflakes",
]
ignore = [
    "B907",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D104",
    "D106",
    "D107",
    "E701",
    "I001",
    "I005",
    "W503",
]
select = ["ANN", "B", "C", "D", "E", "F", "I", "T4", "W"]
extend-select = ["B9"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
